pyproject.toml
[project]
name = "1c-tasks"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "fast-bitrix24>=1.8.4",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "websockets>=15.0",
]
